Things I have done: Finished the /addTime command. The user can add the same day/time multiple times. But this has no effect on the database now. Fixed the user being able to add the same game multiple times.
TO DO:
User can still add multiple skills for a game. The method I used to fix this for games and availability didn't work for skills for some reason.
Start working on a /match command. This command will set the user into groups based on the games they play and the times they are available.